Updated to version 1.5
Main parameters are now in a specific JSON file (param.json) and not in the javascript code. Start images and metadata URL are now in this parameter file (they were previously in donnees.json). Coordinates used in URL and permalink are now in geographic system and in decimal degrees (and not in Mercator coodinates anymore). Colors of the different elements are now only defined in the CSS file and no more in the main.js. Start images can be defined either by their order in the donnees.json file or by the internal name of the layer (used now in permalinks). Add a zoom to extent button (button with globe). Bug corrections.